+++ include/gputils/rand_utils.hpp
@@ -22,10 +22,10 @@ extern std::mt19937 default_rng;
 // -------------------------------------------------------------------------------------------------
 
 
-inline ssize_t rand_int(ssize_t lo, ssize_t hi, std::mt19937 &rng = default_rng)
+inline long rand_int(long lo, long hi, std::mt19937 &rng = default_rng)
 {
     assert(lo < hi);
-    return std::uniform_int_distribution<ssize_t>(lo,hi-1)(rng);   // note hi-1 here!
+    return std::uniform_int_distribution<long>(lo,hi-1)(rng);   // note hi-1 here!
 }
 
 
@@ -40,28 +40,28 @@ inline float rand_uniform(float lo=0.0, float hi=1.0, std::mt19937 &rng = defaul
 
 // Version of randomize() for floating-point types
 template<typename T>
-inline void randomize_f(T *buf, ssize_t nelts, std::mt19937 &rng = default_rng)
+inline void randomize_f(T *buf, long nelts, std::mt19937 &rng = default_rng)
 {
     static_assert(std::is_floating_point_v<T>);
 
     auto dist = std::uniform_real_distribution<T>(-1.0, 1.0);    
-    for (ssize_t i = 0; i < nelts; i++)
+    for (long i = 0; i < nelts; i++)
 	buf[i] = dist(rng);
 }
 
 
 // Version of randomize() for integral types
 template<typename T>
-inline void randomize_i(T *buf, ssize_t nelts, std::mt19937 &rng = default_rng)
+inline void randomize_i(T *buf, long nelts, std::mt19937 &rng = default_rng)
 {
     static_assert(std::is_integral_v<T>);
 
-    ssize_t nbytes = nelts * sizeof(T);
-    ssize_t nints = nbytes / sizeof(int);
+    long nbytes = nelts * sizeof(T);
+    long nints = nbytes / sizeof(int);
     
-    for (ssize_t i = 0; i < nints; i++)
+    for (long i = 0; i < nints; i++)
 	((int *)buf)[i] = rng();
-    for (ssize_t i = nints * sizeof(int); i < nbytes; i++)
+    for (long i = nints * sizeof(int); i < nbytes; i++)
 	((char *)buf)[i] = rng();
 }
 
@@ -69,7 +69,7 @@ inline void randomize_i(T *buf, ssize_t nelts, std::mt19937 &rng = default_rng)
 // General randomize() template, for built-in C++ int/float types.
 // For CUDA __half and __half2, we need specializations (see below).
 template<typename T>
-inline void randomize(T *buf, ssize_t nelts, std::mt19937 &rng = default_rng)
+inline void randomize(T *buf, long nelts, std::mt19937 &rng = default_rng)
 {
     assert(nelts >= 0);
 
@@ -88,24 +88,24 @@ inline void randomize(T *buf, ssize_t nelts, std::mt19937 &rng = default_rng)
 
 // __half randomize() template specialization.
 template<>
-inline void randomize(__half *buf, ssize_t nelts, std::mt19937 &rng)
+inline void randomize(__half *buf, long nelts, std::mt19937 &rng)
 {
     assert(nelts >= 0);    
     auto dist = std::uniform_real_distribution<float>(-1.0f, 1.0f);
     
-    for (ssize_t i = 0; i < nelts; i++)
+    for (long i = 0; i < nelts; i++)
 	buf[i] = __float2half_rn(dist(rng));
 }
 
 
 // __half2 randomize() template specialization.
 template<>
-inline void randomize(__half2 *buf, ssize_t nelts, std::mt19937 &rng)
+inline void randomize(__half2 *buf, long nelts, std::mt19937 &rng)
 {
     assert(nelts >= 0);
     auto dist = std::uniform_real_distribution<float>(-1.0f, 1.0f);
     
-    for (ssize_t i = 0; i < nelts; i++) {
+    for (long i = 0; i < nelts; i++) {
 	float x = dist(rng);
 	float y = dist(rng);
 	buf[i] = __floats2half2_rn(x,y);
@@ -119,20 +119,20 @@ inline void randomize(__half2 *buf, ssize_t nelts, std::mt19937 &rng)
 template<typename T>
 inline void randomly_permute(std::vector<T> &v, std::mt19937 &rng = default_rng)
 {
-    for (size_t i = 1; i < v.size(); i++) {
-	size_t j = rand_int(0, i+1, rng);
+    for (ulong i = 1; i < v.size(); i++) {
+	long j = rand_int(0, i+1, rng);
 	std::swap(v[i], v[j]);
     }
 }
 
 
 // Returns a random permutation of {0,1,...,(n-1)}
-inline std::vector<ssize_t> rand_permutation(ssize_t nelts, std::mt19937 &rng = default_rng)
+inline std::vector<long> rand_permutation(long nelts, std::mt19937 &rng = default_rng)
 {
     assert(nelts >= 0);
     
-    std::vector<ssize_t> v(nelts);
-    for (ssize_t i = 0; i < nelts; i++)
+    std::vector<long> v(nelts);
+    for (long i = 0; i < nelts; i++)
 	v[i] = i;
 
     randomly_permute(v, rng);
@@ -160,7 +160,7 @@ inline T rand_element(const std::initializer_list<T> v, std::mt19937 &rng = defa
 extern std::vector<double> random_doubles_with_fixed_sum(int nelts, double sum);
 
 // Useful in unit tests, when generating randomly-sized arrays.
-extern std::vector<ssize_t> random_integers_with_bounded_product(int nelts, ssize_t bound);
+extern std::vector<long> random_integers_with_bounded_product(int nelts, long bound);
 
     
 } // namespace gputils
