+++ src_lib/rand_utils.cu
@@ -31,7 +31,7 @@ vector<double> random_doubles_with_fixed_sum(int n, double sum)
 }
 
 
-vector<ssize_t> random_integers_with_bounded_product(int n, ssize_t bound)
+vector<long> random_integers_with_bounded_product(int n, long bound)
 {    
     if (n < 1)
 	throw runtime_error("random_integers_with_bounded_product(): 'n' argument must be >= 1");
@@ -41,26 +41,26 @@ vector<ssize_t> random_integers_with_bounded_product(int n, ssize_t bound)
     double target_log = log(rand_uniform(1.01, bound+0.99));
     vector<double> x = random_doubles_with_fixed_sum(n, target_log);
     
-    vector<ssize_t> ret(n);    
+    vector<long> ret(n);    
     for (int i = 0; i < n; i++)
-	ret[i] = ssize_t(exp(x[i]) + 0.5);
+	ret[i] = long(exp(x[i]) + 0.5);
     
     int i0 = rand_int(0, n);
 
     for (int i1 = 0; i1 < n; i1++) {
 	int i = (i0 + i1) % n;
 	
-	ssize_t p = 1;
+	long p = 1;
 	for (int j = 0; j < n; j++)
 	    if (j != i)
 		p *= ret[j];
 
-	ret[i] = ssize_t(exp(target_log-log(p)) + 0.5);
+	ret[i] = long(exp(target_log-log(p)) + 0.5);
 	ret[i] = min(ret[i], bound/p);
 	ret[i] = max(ret[i], 1L);
     }
 
-    ssize_t p = 1;
+    long p = 1;
     for (int j = 0; j < n; j++)
 	p *= ret[j];
